refactor: BaseEntity 추가
