+++ src/main/java/gift/member/model/Member.java
@@ -1,23 +1,12 @@
 package gift.member.model;
 
+import gift.common.model.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import java.time.LocalDateTime;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-@EntityListeners(AuditingEntityListener.class)
 @Entity
-public class Member {
+public class Member extends BaseEntity {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
     @Column(name = "email", nullable = false, unique = true)
     private String email;
     @Column(name = "password", nullable = false)
@@ -26,10 +15,6 @@ public class Member {
     private String name;
     @Column(name = "role")
     private String role;
-    @CreatedDate
-    private LocalDateTime createdDate;
-    @LastModifiedDate
-    private LocalDateTime modifiedDate;
 
     protected Member() {
 
@@ -43,14 +28,10 @@ public class Member {
     }
 
     public Member(Long id, String email, String name, String role) {
-        this.id = id;
+        this.setId(id);
         this.email = email;
         this.name = name;
         this.role = role;
-    }
-
-    public Long getId() {
-        return id;
     }
 
     public String getEmail() {
